Options for emtpy output files, fixes
In first script sections (ll. 40 - 62) adding check for existance of
input data file. If no file exists, a warning is issued and an empty
output file is created. This is needed in order to work with the new
DSMACC make process. Otherwise, script runs as usual.
Furthermore, a new option is added for the data file argument. If "-"
or "0" is passed on to the script, an empty file emissions file is
created deliberately.
New sample output.
FIX: Print of fixed variable EMISS in the output kpp file is now omitted
     as this command is already part of inorganic.kpp.