Options to create empty deposition files, fixes
In first section of makedepos.pl, new if conditions to allow the 
creation of empty KPP emission files in accord to makeemiss.pl.
Initialisation of writefile moved to the header.
In makeemiss.pl, formatting screen output and output in empty KPP file.
Formatting reaction labels of the inorganic KPP file with 'I' before the
reaction number to distinguish inorganic reactions from organic
reactions.
FIX: In if condition on l. 40, change -e option to -f to check for
     existance of file (not file or folder).