Omitting modules in script
In makedepos.pl and makeemiss.pl, deleting the use of MoreUtils.
In makedepos.pl, making varibale $idx an array @idx. Substituting lines
using MoreUtils to find first index in array with grep command 
(ll. 116, 152, and 158). In the following code, the last entry of @idx
is used.
Same changes in makeemiss.pl at ll. 119 - 122.