Additional check for empty kpp output files, fixes
At the end of script makedepos.pl and makeemiss.pl, additional check for
empty output files: If the number of matching species is 0, a warning 
will be issued and the empty sample output file will be created.
Printing additional information on sample kpp file name, when script is
stopped prematurely.
FIXES: Several missing semicolons added.