Combined functions for emiss/depos in wrtkpp, fix
In wrtkpp, combining emikpp and depkpp to wrtmech and write_emi/
write_dep to write_kpp.
Adding on-screen warnings, which species are omitted in the current
(due to missing definitions in the current kpp mechanim files).
Updated doc strings, comments in code, and README.
FIX: Use of deepcopy to save matrix to another matrix to avoid changed 
     values in second matrix, when first matrix changes.